Struct rusoto_support::AWSSupportClient

pub struct AWSSupportClient { /* fields omitted */ }

A client for the AWS Support API.

Methods

impl AWSSupportClient

pub fn new(region: Region) -> AWSSupportClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> AWSSupportClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,