Struct rusoto_support::CaseDetails

pub struct CaseDetails {
    pub case_id: Option<String>,
    pub category_code: Option<String>,
    pub cc_email_addresses: Option<Vec<String>>,
    pub display_id: Option<String>,
    pub language: Option<String>,
    pub recent_communications: Option<RecentCaseCommunications>,
    pub service_code: Option<String>,
    pub severity_code: Option<String>,
    pub status: Option<String>,
    pub subject: Option<String>,
    pub submitted_by: Option<String>,
    pub time_created: Option<String>,
}

A JSON-formatted object that contains the metadata for a support case. It is contained the response from a DescribeCases request. CaseDetails contains the following fields:

caseId. The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

categoryCode. The category of problem for the AWS Support case. Corresponds to the CategoryCode values returned by a call to DescribeServices.

displayId. The identifier for the case on pages in the AWS Support Center.

language. The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

recentCommunications. One or more Communication objects. Fields of these objects are attachments, body, caseId, submittedBy, and timeCreated.

nextToken. A resumption point for pagination.

serviceCode. The identifier for the AWS service that corresponds to the service code defined in the call to DescribeServices.

severityCode. The severity code assigned to the case. Contains one of the values returned by the call to DescribeSeverityLevels.

status. The status of the case in the AWS Support Center.

subject. The subject line of the case.

submittedBy. The email address of the account that submitted the case.

timeCreated. The time the case was created, in ISO-8601 format.

Fields
case_id: Option<String>

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

category_code: Option<String>

The category of problem for the AWS Support case.

cc_email_addresses: Option<Vec<String>>

The email addresses that receive copies of communication about the case.

display_id: Option<String>

The ID displayed for the case in the AWS Support Center. This is a numeric string.

language: Option<String>

The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

recent_communications: Option<RecentCaseCommunications>

The five most recent communications between you and AWS Support Center, including the IDs of any attachments to the communications. Also includes a nextToken that you can use to retrieve earlier communications.

service_code: Option<String>

The code for the AWS service returned by the call to DescribeServices.

severity_code: Option<String>

The code for the severity level returned by the call to DescribeSeverityLevels.

status: Option<String>

The status of the case.

subject: Option<String>

The subject line for the case in the AWS Support Center.

submitted_by: Option<String>

The email address of the account that submitted the case.

time_created: Option<String>

The time that the case was case created in the AWS Support Center.