Struct rusoto_support::Communication

pub struct Communication {
    pub attachment_set: Option<Vec<AttachmentDetails>>,
    pub body: Option<String>,
    pub case_id: Option<String>,
    pub submitted_by: Option<String>,
    pub time_created: Option<String>,
}

A communication associated with an AWS Support case. The communication consists of the case ID, the message body, attachment information, the account email address, and the date and time of the communication.

Fields
attachment_set: Option<Vec<AttachmentDetails>>

Information about the attachments to the case communication.

body: Option<String>

The text of the communication between the customer and AWS Support.

case_id: Option<String>

The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47

submitted_by: Option<String>

The email address of the account that submitted the AWS Support case.

time_created: Option<String>

The time the communication was created.