Crate rusoto_support

AWS Support
The AWS Support API reference is intended for programmers who need detailed information about the AWS Support operations and data types. This service enables you to manage your AWS Support cases programmatically. It uses HTTP methods that return results in JSON format.

The AWS Support service also exposes a set of Trusted Advisor features. You can retrieve a list of checks and their descriptions, get check results, specify checks to refresh, and get the refresh status of checks.

The following list describes the AWS Support case management operations:

Service names, issue categories, and available severity levels. The DescribeServices and DescribeSeverityLevels operations return AWS service names, service codes, service categories, and problem severity levels. You use these values when you call the CreateCase operation.

Case creation, case details, and case resolution. The CreateCase, DescribeCases, DescribeAttachment, and ResolveCase operations create AWS Support cases, retrieve information about cases, and resolve cases.

Case communication. The DescribeCommunications, AddCommunicationToCase, and AddAttachmentsToSet operations retrieve and add communications and attachments to AWS Support cases.

The following list describes the operations available from the AWS Support service for Trusted Advisor:

DescribeTrustedAdvisorChecks returns the list of checks that run against your AWS resources.

Using the checkId for a specific check returned by DescribeTrustedAdvisorChecks, you can call DescribeTrustedAdvisorCheckResult to obtain the results for the check you specified.

DescribeTrustedAdvisorCheckSummaries returns summarized results for one or more Trusted Advisor checks.

RefreshTrustedAdvisorCheck requests that Trusted Advisor rerun a specified check.

DescribeTrustedAdvisorCheckRefreshStatuses reports the refresh status of one or more checks.

For authentication of requests, AWS Support uses Signature Version 4 Signing Process.

See About the AWS Support API in the AWS Support User Guide for information about how to use this service to create and manage your support cases, and how to call Trusted Advisor for results of checks on your resources.

If you're using the service, you're probably looking for AWSSupportClient and AWSSupport.

Structs

AWSSupportClient
A client for the AWS Support API.

AddAttachmentsToSetRequest

AddAttachmentsToSetResponse
The ID and expiry time of the attachment set returned by the AddAttachmentsToSet operation.

AddCommunicationToCaseRequest
To be written.

AddCommunicationToCaseResponse
The result of the AddCommunicationToCase operation.

Attachment
An attachment to a case communication. The attachment consists of the file name and the content of the file.

AttachmentDetails
The file name and ID of an attachment to a case communication. You can use the ID to retrieve the attachment with the DescribeAttachment operation.

CaseDetails

A JSON-formatted object that contains the metadata for a support case. It is contained the response from a DescribeCases request. CaseDetails contains the following fields:

caseId. The AWS Support case ID requested or returned in the call. The case ID is an alphanumeric string formatted as shown in this example: case-12345678910-2013-c4c1d2bf33c5cf47.

categoryCode. The category of problem for the AWS Support case. Corresponds to the CategoryCode values returned by a call to DescribeServices.

displayId. The identifier for the case on pages in the AWS Support Center.

language. The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.

recentCommunications. One or more Communication objects. Fields of these objects are attachments, body, caseId, submittedBy, and timeCreated.

nextToken. A resumption point for pagination.

serviceCode. The identifier for the AWS service that corresponds to the service code defined in the call to DescribeServices.

severityCode. The severity code assigned to the case. Contains one of the values returned by the call to DescribeSeverityLevels.

status. The status of the case in the AWS Support Center.

subject. The subject line of the case.

submittedBy. The email address of the account that submitted the case.

timeCreated. The time the case was created, in ISO-8601 format.

Category
A JSON-formatted name/value pair that represents the category name and category code of the problem, selected from the DescribeServices response for each AWS service.

Communication
A communication associated with an AWS Support case. The communication consists of the case ID, the message body, attachment information, the account email address, and the date and time of the communication.

CreateCaseRequest

CreateCaseResponse
The AWS Support case ID returned by a successful completion of the CreateCase operation.

DescribeAttachmentRequest

DescribeAttachmentResponse
The content and file name of the attachment returned by the DescribeAttachment operation.

DescribeCasesRequest

DescribeCasesResponse
Returns an array of CaseDetails objects and a nextToken that defines a point for pagination in the result set.

DescribeCommunicationsRequest

DescribeCommunicationsResponse
The communications returned by the DescribeCommunications operation.

DescribeServicesRequest

DescribeServicesResponse
The list of AWS services returned by the DescribeServices operation.

DescribeSeverityLevelsRequest

DescribeSeverityLevelsResponse
The list of severity levels returned by the DescribeSeverityLevels operation.

DescribeTrustedAdvisorCheckRefreshStatusesRequest

DescribeTrustedAdvisorCheckRefreshStatusesResponse
The statuses of the Trusted Advisor checks returned by the DescribeTrustedAdvisorCheckRefreshStatuses operation.

DescribeTrustedAdvisorCheckResultRequest

DescribeTrustedAdvisorCheckResultResponse
The result of the Trusted Advisor check returned by the DescribeTrustedAdvisorCheckResult operation.

DescribeTrustedAdvisorCheckSummariesRequest

DescribeTrustedAdvisorCheckSummariesResponse
The summaries of the Trusted Advisor checks returned by the DescribeTrustedAdvisorCheckSummaries operation.

DescribeTrustedAdvisorChecksRequest

DescribeTrustedAdvisorChecksResponse
Information about the Trusted Advisor checks returned by the DescribeTrustedAdvisorChecks operation.

RecentCaseCommunications
The five most recent communications associated with the case.

RefreshTrustedAdvisorCheckRequest

RefreshTrustedAdvisorCheckResponse
The current refresh status of a Trusted Advisor check.

ResolveCaseRequest

ResolveCaseResponse
The status of the case returned by the ResolveCase operation.

Service
Information about an AWS service returned by the DescribeServices operation.

SeverityLevel
A code and name pair that represent a severity level that can be applied to a support case.

TrustedAdvisorCategorySpecificSummary
The container for summary information that relates to the category of the Trusted Advisor check.

TrustedAdvisorCheckDescription
The description and metadata for a Trusted Advisor check.

TrustedAdvisorCheckRefreshStatus
The refresh status of a Trusted Advisor check.

TrustedAdvisorCheckResult
The results of a Trusted Advisor check returned by DescribeTrustedAdvisorCheckResult.

TrustedAdvisorCheckSummary
A summary of a Trusted Advisor check result, including the alert status, last refresh, and number of resources examined.

TrustedAdvisorCostOptimizingSummary
The estimated cost savings that might be realized if the recommended actions are taken.

TrustedAdvisorResourceDetail
Contains information about a resource identified by a Trusted Advisor check.

TrustedAdvisorResourcesSummary
Details about AWS resources that were analyzed in a call to Trusted Advisor DescribeTrustedAdvisorCheckSummaries.

Enums

AddAttachmentsToSetError
Errors returned by AddAttachmentsToSet

AddCommunicationToCaseError
Errors returned by AddCommunicationToCase

CreateCaseError
Errors returned by CreateCase

DescribeAttachmentError
Errors returned by DescribeAttachment

DescribeCasesError
Errors returned by DescribeCases

DescribeCommunicationsError
Errors returned by DescribeCommunications

DescribeServicesError
Errors returned by DescribeServices

DescribeSeverityLevelsError
Errors returned by DescribeSeverityLevels

DescribeTrustedAdvisorCheckRefreshStatusesError
Errors returned by DescribeTrustedAdvisorCheckRefreshStatuses

DescribeTrustedAdvisorCheckResultError
Errors returned by DescribeTrustedAdvisorCheckResult

DescribeTrustedAdvisorCheckSummariesError
Errors returned by DescribeTrustedAdvisorCheckSummaries

DescribeTrustedAdvisorChecksError
Errors returned by DescribeTrustedAdvisorChecks

RefreshTrustedAdvisorCheckError
Errors returned by RefreshTrustedAdvisorCheck

ResolveCaseError
Errors returned by ResolveCase

Traits

AWSSupport
Trait representing the capabilities of the AWS Support API. AWS Support clients implement this trait.